requirements.txt
asgiref==3.3.1
Django==3.1.6
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
phonenumberslite==8.12.18
pytz==2021.1
sqlparse==0.4.1
